Move some constants from types to constants files. Split constants into multiple files so it can be imported better
