+++ src/types.ts
@@ -1,28 +1,5 @@
 import { Pool as SDKPool, SwapV2, Token as SDKToken } from '@balancer-labs/sdk';
 
-export const Network = {
-    MAINNET: 1,
-    GOERLI: 5,
-    KOVAN: 42,
-    POLYGON: 137,
-    ARBITRUM: 42161
-}
-
-export const NativeAssetAddress = {
-    ETH: "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE",
-    MATIC: "0x7d1afa7b718fb893db30a3abc0cfc608aacfebb0"
-}
-
-export const NativeAssetId = {
-    ETH: "ethereum",
-    MATIC: "matic-network"
-}
-
-export const NativeAssetPriceSymbol = {
-    ETH: "eth",
-    MATIC: "matic"
-}
-
 export interface Order {
     sellToken: string;
     buyToken: string;
